// servers/src/models/address.model.ts
export type Commune = {
    code: string;
    name: string;
    englishName: string;
    administrativeLevel: string;
    provinceCode: string;
    provinceName: string;
    decree: string;
}

export type Province = {
    code: string;
    name: string;
    englishName: string;
    administrativeLevel: string;
    decree: string;
}